Capitalization fixes - rename files and folders
