+++ health-on-time/src/Pages/HomePage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import LandingPage from "../components/LandingPage";
+import LandingPage from "../Components/LandingPage";
 
 const HomePage = () => {
   return (
